+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+Fixed vertex and fragments locations for `BunchFilterConfiguration`
+
+## 0.0.2
+
+Fixed default path for `BunchFilterConfiguration`
+
 ## 0.0.1
 
 Initial Version of the library.